Changed: Cleaned up Cappers
Cappers now have the abstract function to get all online available
players. And getAvailablePlayers in the main abstract class for code
reuse. This function can be overwritten if nessesary.
