+++ src/subside/plugins/koth/adapter/captypes/CappingFactionUUID.java
@@ -1,6 +1,7 @@
 package subside.plugins.koth.adapter.captypes;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
@@ -69,26 +70,12 @@ public class CappingFactionUUID extends CappingGroup {
         }
         return false;
     }
+
     
-    @Override
-    public List<Player> getAvailablePlayers(Capable area){
-        List<Player> list = new ArrayList<Player>();
-        
-        List<Player> players = faction.getOnlinePlayers();
-        if(players.size() > 0){
-            for(Player player : players){
-                if(area.isInArea(player)){
-                    list.add(player);
-                }
-            }
-            
-            if(list.size() < 1){
-                list.add(players.get(0));
-            }
-        }
-        return list;
+    @Override 
+    public List<Player> getAllOnlinePlayers(){
+        return faction.getOnlinePlayers();
     }
-    
 
     public static Capper getFromUniqueName(String name){
         return new CappingFactionUUID(Factions.getInstance().getFactionById(name));
